requirements.txt
Flask
pandas
bokeh==1.4.0
gunicorn
folium
h3
json
numpy
math
branca.colormap
geojson.feature
